refactor: rename 'service' file to more suitable editReleaseDataService
All service files in the admin project are currently called `service.ts`
but this reduces discoverability and is hard to use when importing
multiple services in the same file.

In the future, we should progressively rename these to more accurately
describe the service they actually represent.
