+++ src/explore-education-statistics-admin/src/pages/release/edit-release/manage-datablocks/components/ChartBuilder.tsx
@@ -8,7 +8,7 @@ import ChartDataSelector, {
 } from '@admin/pages/release/edit-release/manage-datablocks/components/ChartDataSelector';
 import ChartTypeSelector from '@admin/pages/release/edit-release/manage-datablocks/components/ChartTypeSelector';
 import styles from '@admin/pages/release/edit-release/manage-datablocks/components/graph-builder.module.scss';
-import service from '@admin/services/release/edit-release/data/service';
+import editReleaseDataService from '@admin/services/release/edit-release/data/editReleaseDataService';
 import withErrorControl, {
   ErrorControlProps,
 } from '@admin/validation/withErrorControl';
@@ -249,7 +249,7 @@ const ChartBuilder = ({
           },
         },
         labels: chartLabels,
-        chartFileDownloadService: service.downloadChartFile,
+        chartFileDownloadService: editReleaseDataService.downloadChartFile,
 
         ...chartOptions,
       });
