+++ src/explore-education-statistics-admin/src/services/release/edit-release/data/editReleaseDataService.ts
@@ -1,14 +1,5 @@
 import client from '@admin/services/util/service';
 
-import {
-  AncillaryFile,
-  ChartFile,
-  DataFile,
-  UploadAncillaryFileRequest,
-  UploadChartFileRequest,
-  UploadDataFilesRequest,
-} from './types';
-
 interface GetFileResponse {
   extension: string;
   name: string;
@@ -44,11 +35,59 @@ const downloadFile = (blob: Blob, fileName: string) => {
   link.remove();
 };
 
+export interface DataFile {
+  title: string;
+  filename: string;
+  fileSize: {
+    size: number;
+    unit: string;
+  };
+  rows: number;
+  metadataFilename: string;
+  userName: string;
+  created: Date;
+  canDelete?: boolean;
+}
+
+export interface UploadDataFilesRequest {
+  subjectTitle: string;
+  dataFile: File;
+  metadataFile: File;
+}
+
+export interface AncillaryFile {
+  title: string;
+  filename: string;
+  fileSize: {
+    size: number;
+    unit: string;
+  };
+}
+
+export interface UploadAncillaryFileRequest {
+  name: string;
+  file: File;
+}
+
+export interface ChartFile {
+  title: string;
+  filename: string;
+  fileSize: {
+    size: number;
+    unit: string;
+  };
+}
+
+export interface UploadChartFileRequest {
+  name: string;
+  file: File;
+}
+
 /**
  * A temporary step to provide a row count to the front end whilst it does not yet exist in the API.
  */
 
-const service = {
+const editReleaseDataService = {
   getReleaseDataFiles(releaseId: string): Promise<DataFile[]> {
     return client
       .get<GetFileResponse[]>(`/release/${releaseId}/data`)
@@ -193,4 +232,4 @@ const service = {
   },
 };
 
-export default service;
+export default editReleaseDataService;
